Disable doclint in Java 8 builds
Doclint is an exceptionally new feature in Java 8 that causes javadoc
generation to produce an error any time it encounters invalid javadoc
syntax. Unfortunatly, if your javadocs are not properly doclint-compliant
(like ours), this will cause generating javadocs, and anything that depends
on it, i.e. `mvn install` to fail with a large number of javadoc errors.
Unless we want to write a very large amount of javadoc information, we
should disable doclint for now. This change is set up so that it will
add the flag to disable doclint, only for Java 8, for Java 7, this will
allow build to continue without any problems.